+++ packages/react-kit/src/components/modal/components/Commit/CommitSuccess/CommitSuccess.tsx
@@ -115,7 +115,7 @@ export function CommitSuccess({ onHouseClick, exchangeId, commitHash }: Props) {
                     }}
                   >
                     View on{" "}
-                    {[80001, 137].includes(config.chainId)
+                    {[80002, 137].includes(config.chainId)
                       ? "Polygonscan"
                       : "Etherscan"}{" "}
                     <ArrowSquareUpRight