+++ packages/react-kit/src/components/widgets/commit/CommitModalWithOffer.tsx
@@ -82,7 +82,7 @@ function WithProductOrOffer(
                   ({
                     variations: productV1Item.variations,
                     offer: bundle.offer
-                  } as VariantV1)
+                  }) as VariantV1
               );
             })
             .filter(isTruthy);