+++ packages/react-kit/src/lib/opensea/getOpenSeaUrl.ts
@@ -4,12 +4,12 @@ import { getExchangeTokenId } from "../utils/exchange";
 
 const openSeaUrlMap = new Map([
   [
-    "testing", // Mumbai
+    "testing", // testnets
     new Map([
       [
-        "testing-80001-0",
+        "testing-80002-0",
         (tokenId: string, contractAddress: string) =>
-          `https://testnets.opensea.io/assets/mumbai/${contractAddress}/${tokenId}`
+          `https://testnets.opensea.io/assets/amoy/${contractAddress}/${tokenId}` // TO BE CONFIRMED
       ],
       [
         "testing-11155111-0",
@@ -19,12 +19,12 @@ const openSeaUrlMap = new Map([
     ])
   ],
   [
-    "staging", // Mumbai
+    "staging", // testnets
     new Map([
       [
-        "staging-80001-0",
+        "staging-80002-0",
         (tokenId: string, contractAddress: string) =>
-          `https://testnets.opensea.io/assets/mumbai/${contractAddress}/${tokenId}`
+          `https://testnets.opensea.io/assets/amoy/${contractAddress}/${tokenId}` // TO BE CONFIRMED
       ],
       [
         "staging-11155111-0",