+++ packages/react-kit/src/stories/cta/dispute/RaiseDisputeButton.stories.tsx
@@ -41,7 +41,7 @@ export const WithExtraInfo: ComponentStory<typeof RaiseDisputeButton> =
 
 Simple.args = {
   coreSdkConfig: {
-    configId: "testing-80001-0",
+    configId: "testing-80002-0",
     envName: "testing"
     // metaTransactionsApiKey: undefined,
     // metaTransactionsApiId: "dummyApiId",
@@ -67,7 +67,7 @@ Simple.args = {
 
 WithExtraInfo.args = {
   coreSdkConfig: {
-    configId: "testing-80001-0",
+    configId: "testing-80002-0",
     envName: "testing"
     // metaTransactionsApiKey: undefined,
   },