+++ packages/react-kit/src/stories/cta/dispute/RemoveFeesDisputeResolverButton.stories.tsx
@@ -40,7 +40,7 @@ export const WithExtraInfo: ComponentStory<typeof RemoveFeesDisputeResolver> =
 
 Simple.args = {
   coreSdkConfig: {
-    configId: "testing-80001-0",
+    configId: "testing-80002-0",
     envName: "testing",
     web3Provider: undefined
   },
@@ -66,7 +66,7 @@ Simple.args = {
 
 WithExtraInfo.args = {
   coreSdkConfig: {
-    configId: "testing-80001-0",
+    configId: "testing-80002-0",
     envName: "testing",
     web3Provider: undefined
   },