+++ packages/react-kit/src/stories/cta/exchange/CompleteButton.stories.tsx
@@ -39,7 +39,7 @@ export const WithExtraInfo: ComponentStory<typeof CompleteButton> =
 
 Simple.args = {
   coreSdkConfig: {
-    configId: "testing-80001-0",
+    configId: "testing-80002-0",
     envName: "testing"
   },
   exchangeId: "28",
@@ -63,7 +63,7 @@ Simple.args = {
 
 WithExtraInfo.args = {
   coreSdkConfig: {
-    configId: "testing-80001-0",
+    configId: "testing-80002-0",
     envName: "testing"
   },
   exchangeId: "28",