+++ scripts/change-contract-uri.ts
@@ -13,7 +13,7 @@ program
   )
   .argument("<CONTRACT_URI>", "Contract URI to set in the seller's voucher")
   .option("-e, --env <ENV_NAME>", "Target environment", "testing")
-  .option("-c, --configId <CONFIG_ID>", "Config id", "testing-80001-0")
+  .option("-c, --configId <CONFIG_ID>", "Config id", "testing-80002-0")
   .parse(process.argv);
 
 async function main() {
@@ -21,7 +21,7 @@ async function main() {
 
   const opts = program.opts();
   const envName = opts.env || "testing";
-  const configId = opts.configId || "testing-80001-0";
+  const configId = opts.configId || "testing-80002-0";
   const defaultConfig = getEnvConfigById(envName as EnvironmentType, configId);
   const chainId = defaultConfig.chainId;
   const sellerWallet = new Wallet(sellerPrivateKey);