+++ scripts/complete-exchange.ts
@@ -23,7 +23,7 @@ program
     "Private key of the account issuing the transaction."
   )
   .option("-e, --env <ENV_NAME>", "Target environment", "testing")
-  .option("-c, --configId <CONFIG_ID>", "Config id", "testing-80001-0")
+  .option("-c, --configId <CONFIG_ID>", "Config id", "testing-80002-0")
   .option(
     "-m, --metaTx <META_TX_CONFIG>",
     "JSON file with the meta tx parameters"
@@ -44,7 +44,7 @@ async function main() {
 
   const opts = program.opts();
   const envName = (opts.env as EnvironmentType) || "testing";
-  const configId = opts.configId || "testing-80001-0";
+  const configId = opts.configId || "testing-80002-0";
   const defaultConfig = getEnvConfigById(envName, configId);
   const wallet = new Wallet(privateKey);
   let metaTx: Partial<MetaTxConfig> | undefined = undefined;