+++ scripts/create-offer-batch.ts
@@ -16,7 +16,7 @@ program
   )
   .argument("<OFFER_DATA>", "JSON file with the Offers specific parameters")
   .option("-e, --env <ENV_NAME>", "Target environment", "testing")
-  .option("-c, --configId <CONFIG_ID>", "Config id", "testing-80001-0")
+  .option("-c, --configId <CONFIG_ID>", "Config id", "testing-80002-0")
   .option(
     "--template <TEMPLATE>",
     "JSON file with the Offers common parameters"
@@ -28,7 +28,7 @@ async function main() {
 
   const opts = program.opts();
   const envName = opts.env || "testing";
-  const configId = opts.configId || "testing-80001-0";
+  const configId = opts.configId || "testing-80002-0";
   const defaultConfig = getEnvConfigById(envName as EnvironmentType, configId);
   const chainId = defaultConfig.chainId;
   const offersRawData = fs.readFileSync(offerDataJsonFile);