+++ scripts/get-seller.ts
@@ -10,7 +10,7 @@ program
   .description("Get a Seller.")
   .argument("<SELLER_ADDRESS>", "Address of the Seller to look after.")
   .option("-e, --env <ENV_NAME>", "Target environment", "testing")
-  .option("-c, --configId <CONFIG_ID>", "Config id", "testing-80001-0")
+  .option("-c, --configId <CONFIG_ID>", "Config id", "testing-80002-0")
   .parse(process.argv);
 
 async function main() {
@@ -18,7 +18,7 @@ async function main() {
 
   const opts = program.opts();
   const envName = opts.env || "testing";
-  const configId = opts.configId || "testing-80001-0";
+  const configId = opts.configId || "testing-80002-0";
   const defaultConfig = getEnvConfigById(envName as EnvironmentType, configId);
 
   const coreSDK = CoreSDK.fromDefaultConfig({