+++ packages/react-kit/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.29.1](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/react-kit@0.29.0...@bosonprotocol/react-kit@0.29.1) (2024-04-15)
+
+**Note:** Version bump only for package @bosonprotocol/react-kit
+
+
+
+
+
 # [0.29.0](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/react-kit@0.28.1...@bosonprotocol/react-kit@0.29.0) (2024-04-05)